+++ NEWS.md
@@ -1,3 +1,7 @@
+# brmsmargins devel
+
+* Fixed a bug preventing predictions integrating out random effects for mixed effects models with a random intercept only (reported in Issue#1). Thanks to @ajnafa for reporting.
+
 # brmsmargins 0.1.0
 
 * Initial release